# Conflicts:
#	src/main/java/ar/edu/unlam/tallerweb1/delivery/ControladorLogin.java
#	src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosComentario.java
#	src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosMascotas.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/mascotas/Mascota.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/mascotas/ServicioMascota.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/mascotas/ServicioMascotaImpl.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/ServicioUsuario.java
#	src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/ServicioUsuarioImpl.java
#	src/main/java/ar/edu/unlam/tallerweb1/infrastructure/RepositorioMascota.java
#	src/main/java/ar/edu/unlam/tallerweb1/infrastructure/RepositorioMascotaImpl.java
#	src/main/java/ar/edu/unlam/tallerweb1/infrastructure/RepositorioUsuario.java
#	src/main/java/ar/edu/unlam/tallerweb1/infrastructure/RepositorioUsuarioImpl.java
#	src/main/webapp/WEB-INF/vistas/detalle-mascota.jsp
#	src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorLoginTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorMascotaTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/domain/ServicioComentariosTest.java
#	src/test/java/ar/edu/unlam/tallerweb1/infrastructure/RepositorioMascotaTest.java

+++ src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosComentario.java
@@ -25,11 +25,11 @@ public class DatosComentario {
         this.mensaje = mensaje;
     }
 
-    public int getIdUsuario() {
+    public Long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(int idUsuario) {
+    public void setIdUsuario(Long idUsuario) {
         this.idUsuario = idUsuario;
     }
